Refactor NavbarComp component
Refactored the NavbarComp component to improve code readability and maintainability. Replaced the use of index as the key in the map function with a unique key value from the link object. Also added conditional styling to the Link component based on the current pathname. Removed redundant NavbarItem and Contact link.